refactor(db): shared connection factory + ssl ca resolver
